Fix tag tool custom prop enum getting out of sync when removing props
Example problems:
Having 3 custom props: ("cp_xxxxxx", "A"), "B" and "C".
- If B is selected and A is deleted, the enum would switch to C.
- If C is selected and A is deleted, the enum would error out of range.

The problems happen because, internally, Blender stores the enum value as an
int (1, 2, 3), instead of the "cp_xxxxxx" identifier, due to the enum prop
being a dynamic and non flag.
This makes it unable to keep the selected value stable.

As a workaround, the tag tool now keeps the current enum value as a string in
its own variable and overrides the enum's get/set to keep the string copy in
sync.

Note: having no custom props still triggers Blender internal warnings (harmless)