Remove prototype code to support grouping by any custom property
After some testing, grouping on anything didn't seem that useful, specially for
the case of enum flags that imply that the same thumbnail can show in multiple
groups or not at all.
As soon as the property being used for grouping is not correlated with the timeline
the interaction gets confusing.

Instead of grouping on anything, we'll go with explicit support for sequences/scenes.