Predefine storage dir for thumbnails
Before this change, the user could specify in which folder to store
the thumbnails. This could lead to a bad experience, since upon
thumbnails regeneration the content of a folder would be wiped.

Now the folder is automatically created in the default data directory
used by each OS, and is a read-only property accessible in the user
preferences panel.