Fix blend file being unsaved despite having no changes and other undo bugs
- blend file kept being marked as unsaved because the select and tag tools
react on mouse move to set the hovered thumbnail. Despite this happening on
invole and causing no changes to the blend file data, it was marked as
unsaved. Operators now return 'CANCELLED' to avoid undo stack spam.

- Shot select tool is now undoable.

- While undoing, for reasons unclear to me, Blender will clear the addon's
thumbnail list, or the thumbnails' image data. It doesn't seem to happen
deterministically, but it always does in less than 5 undo steps.
To try and prevent this, I refactored code to use more indexes and less
thumbnail objects (only partially helps?).
In the end, on draw, I just check if Blender deleted data and reload it if
so. This is not ideal, but doesn't seem to noticeable for the user experience.