Conversion to extension for Blender 4.2+
- Added manifest
- Fixed ID for the Blender preferences can not be the package name.
Blender uses the extension prefix, so this add-on must too.

I was undecided if it is nice to have the logging using __name__ ir the package_name or the ADDON_ID.
On one hand it could be nice if it is a single name as the logs would be aligned and it is easier to
programatically change the verbosity level. On the other hand, it's nice to have more qualified names.